+++ .eslintrc.json
@@ -12,6 +12,6 @@
     }
   },
   "extends": ["react-app", "plugin:react/recommended", "plugin:prettier/recommended"],
-  "plugins": ["react", "react-hooks"],
+  //"plugins": ["react", "react-hooks"],
   "rules": {}
 }
