+++ src/_nav.js
@@ -550,11 +550,6 @@ const _nav = [
       },
       {
         component: CNavItem,
-        name: 'Maintenance',
-        to: '/cipp/maintenance',
-      },
-      {
-        component: CNavItem,
         name: 'Documentation',
         href: 'https://cipp.app',
         target: '_blank',
