+++ src/views/cipp/CIPPSettings.js
@@ -36,19 +36,27 @@ import {
   useListExcludedTenantsQuery,
 } from 'src/store/api/tenants'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { faCircleNotch, faTrash, faTrashAlt } from '@fortawesome/free-solid-svg-icons'
+import {
+  faCircleNotch,
+  faLink,
+  faScroll,
+  faTrash,
+  faTrashAlt,
+} from '@fortawesome/free-solid-svg-icons'
 import { useListTenantsQuery } from 'src/store/api/tenants'
 import { useLazyEditDnsConfigQuery, useLazyGetDnsConfigQuery } from 'src/store/api/domains'
 import { useDispatch, useSelector } from 'react-redux'
 import { CippTable } from 'src/components/tables'
 import { CippPage, CippPageList } from 'src/components/layout'
-import { RFFCFormSwitch, RFFCFormInput } from 'src/components/forms'
+import { RFFCFormSwitch, RFFCFormInput, RFFCFormSelect } from 'src/components/forms'
 import { Form } from 'react-final-form'
 import useConfirmModal from 'src/hooks/useConfirmModal'
 import { setCurrentTenant } from 'src/store/features/app'
-import { ModalService, TenantSelectorMultiple } from 'src/components/utilities'
+import { CippCodeBlock, ModalService, TenantSelectorMultiple } from 'src/components/utilities'
 import CippListOffcanvas from 'src/components/utilities/CippListOffcanvas'
 import { TitleButton } from 'src/components/buttons'
+import Skeleton from 'react-loading-skeleton'
+import { Buffer } from 'buffer'
 
 const CIPPSettings = () => {
   const [active, setActive] = useState(1)
@@ -70,6 +78,9 @@ const CIPPSettings = () => {
         <CNavItem active={active === 5} onClick={() => setActive(5)} href="#">
           Licenses
         </CNavItem>
+        <CNavItem active={active === 6} onClick={() => setActive(6)} href="#">
+          Maintenance
+        </CNavItem>
       </CNav>
       <CTabContent>
         <CTabPane visible={active === 1} className="mt-3">
@@ -86,6 +97,9 @@ const CIPPSettings = () => {
         </CTabPane>
         <CTabPane visible={active === 5} className="mt-3">
           <LicenseSettings />
+        </CTabPane>
+        <CTabPane visible={active === 6} className="mt-3">
+          <Maintenance />
         </CTabPane>
       </CTabContent>
     </CippPage>
@@ -880,6 +894,143 @@ const DNSSettings = () => {
           </CCardBody>
         </CCard>
       )}
+    </>
+  )
+}
+const Maintenance = () => {
+  const [selectedScript, setSelectedScript] = useState()
+  const [listBackend, listBackendResult] = useLazyGenericGetRequestQuery()
+  const [listScript, listScriptResult] = useLazyGenericGetRequestQuery()
+  const [listScriptLink, listScriptLinkResult] = useLazyGenericGetRequestQuery()
+
+  const handleSubmit = async (values) => {
+    listScript({ path: 'api/ExecMaintenanceScripts', params: values })
+    setSelectedScript(values.ScriptFile)
+  }
+
+  const handleGetLink = () => {
+    console.log('Making link')
+    listScriptLink({
+      path: 'api/ExecMaintenanceScripts',
+      params: { ScriptFile: selectedScript, MakeLink: 'True' },
+    })
+  }
+  return (
+    <>
+      {listBackendResult.isUninitialized && listBackend({ path: 'api/ExecMaintenanceScripts' })}
+      <CRow>
+        <CCol>
+          <CCard className="options-card">
+            <CCardHeader>
+              <CCardTitle className="d-flex justify-content-between">Maintenance</CCardTitle>
+            </CCardHeader>
+            <CCardBody>
+              <Form
+                initialValues={{}}
+                onSubmit={handleSubmit}
+                render={({ handleSubmit, submitting, values }) => {
+                  return (
+                    <CForm onSubmit={handleSubmit}>
+                      {listBackendResult.isFetching && (
+                        <>
+                          <CRow>
+                            <CCol>
+                              <Skeleton count={5} />
+                            </CCol>
+                          </CRow>
+                        </>
+                      )}
+                      {!listBackendResult.isFetching && listBackendResult.isSuccess && (
+                        <>
+                          <CRow>
+                            <CCol>
+                              <RFFCFormSelect
+                                name="ScriptFile"
+                                label="Script File"
+                                placeholder="-- Select a script --"
+                                values={listBackendResult.data.ScriptFiles}
+                              />
+                            </CCol>
+                          </CRow>
+                          <CRow className="mb-3">
+                            <CCol>
+                              <CButton type="submit" disabled={submitting}>
+                                <FontAwesomeIcon icon={faScroll} className="me-2" />
+                                Load Script
+                              </CButton>
+                            </CCol>
+                          </CRow>
+                        </>
+                      )}
+                    </CForm>
+                  )
+                }}
+              />
+            </CCardBody>
+          </CCard>
+        </CCol>
+      </CRow>
+      <CRow>
+        <CCol>
+          {listScriptResult.isFetching && (
+            <CCard>
+              <CCardBody>
+                <Skeleton count={10} />
+              </CCardBody>
+            </CCard>
+          )}
+          {!listScriptResult.isFetching && listScriptResult.isSuccess && (
+            <CCard>
+              <CCardHeader>
+                <CCardTitle>Script Details</CCardTitle>
+              </CCardHeader>
+              <CCardBody>
+                <p>
+                  <CButton type="submit" onClick={handleGetLink}>
+                    <FontAwesomeIcon icon={faLink} className="me-2" />
+                    Create Link
+                  </CButton>
+                </p>
+                {listScriptLinkResult.isSuccess && (
+                  <p>
+                    {listScriptLinkResult.data.Link !== undefined && (
+                      <>
+                        <p>
+                          Copy this text into a PowerShell terminal, we recommend Azure Cloud Shell.
+                          Azure modules and the az command line utilties are required for these
+                          scripts to work. The link is valid for 5 minutes.
+                        </p>
+                        <CippCodeBlock
+                          language="text"
+                          showLineNumbers={false}
+                          wrapLongLines={true}
+                          code={
+                            'irm ' +
+                            window.location.origin +
+                            listScriptLinkResult.data.Link +
+                            ' | iex'
+                          }
+                        />
+                      </>
+                    )}
+                  </p>
+                )}
+                {listScriptResult.data.ScriptContent !== undefined && (
+                  <p>
+                    <h5>Maintenance Script Contents</h5>
+                    <CippCodeBlock
+                      language="powershell"
+                      showLineNumbers={true}
+                      wrapLongLines={false}
+                      code={Buffer.from(listScriptResult.data.ScriptContent, 'base64').toString()}
+                    />
+                  </p>
+                )}
+              </CCardBody>
+            </CCard>
+          )}
+        </CCol>
+      </CRow>
     </>
   )
 }
